Using variable from airflow UI
