--- airflow/dags/utils/make_folder.py
@@ -1,11 +1,0 @@
-# Get href values for download func
-#########################
-####SUPPORT FUNCTIONS####
-#########################
-def _make_folder(path) -> None:
-    import os
-    
-    if not os.path.exists(path):
-        os.makedirs(path)
-    else:
-        pass
